refactor(infrastructure): name the stub casting for its combination
